acme_project/requirements.txt
absl-py==1.4.0
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiopg==1.4.0
aiosignal==1.3.1
altgraph==0.17.4
annotated-types==0.7.0
anyio==3.6.2
asgiref==3.8.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==24.2.0
auto-py-to-exe==2.44.1
Babel==2.9.1
beautifulsoup4==4.12.3
bottle==0.13.1
bottle-websocket==0.2.9
cachetools==5.3.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.1.0
contourpy==1.3.0
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
Django==5.1.5
django-bootstrap4==24.4
django-bootstrap5==24.3
django-debug-toolbar==5.0.1
Eel==0.17.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
flatbuffers==23.5.26
fonttools==4.54.1
frozenlist==1.4.1
future==1.0.0
gast==0.4.0
gevent==24.2.1
gevent-websocket==0.10.1
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.56.2
h11==0.14.0
h5py==3.9.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
keras==2.13.1
kiwisolver==1.4.7
libclang==16.0.6
magic-filter==1.0.12
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.9.2
multidict==6.1.0
numpy==1.24.3
oauthlib==3.2.2
openpyxl==3.1.5
opt-einsum==3.3.0
outcome==1.3.0.post0
packaging==23.1
pandas==2.2.3
pefile==2024.8.26
pillow==10.4.0
propcache==0.2.0
protobuf==4.24.0
psycopg2-binary==2.9.10
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
pyparsing==3.1.4
PySocks==1.7.1
pyTelegramBotAPI==4.10.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==20.2
pytz==2024.2
pywin32-ctypes==0.2.3
requests==2.28.2
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
selenium==4.26.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.36
sqlparse==0.5.3
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.15
webmoney==0.0.1
websocket-client==1.8.0
Werkzeug==2.3.6
whichcraft==0.6.1
wrapt==1.15.0
wsproto==1.2.0
yarl==1.14.0
zope.event==5.0
zope.interface==7.0.3
